ajustes id entidades

+++ src/main/java/br/com/mattos/simuladorinvestimento/application/dto/request/SimularInvestimentoRequestDTO.java
@@ -4,8 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-/** DTO para requisição de simulação de investimento. */
-public record SimularInvestimentoRequestDTO(@NotNull @Positive Integer clienteId, @NotNull @Positive Double valor,
+public record SimularInvestimentoRequestDTO(@NotNull @Positive Long clienteId, @NotNull @Positive Double valor,
                                             @NotNull @Positive Integer prazoMeses,@NotBlank String tipoProduto) {
 
 }
