+++ src/main/java/br/com/mattos/simuladorinvestimento/application/dto/response/ProdutoResponseDTO.java
@@ -1,6 +1,5 @@
 package br.com.mattos.simuladorinvestimento.application.dto.response;
 
-/** DTO de resposta contendo informações do produto. */
-public record ProdutoResponseDTO(Integer id,String nome,String tipo,Double rentabilidade,String risco) {
+public record ProdutoResponseDTO(Long id,String nome,String tipo,Double rentabilidade,String risco) {
 
 }
