+++ src/main/java/br/com/mattos/simuladorinvestimento/domain/model/Produto.java
@@ -1,6 +1,5 @@
 package br.com.mattos.simuladorinvestimento.domain.model;
 
-/** Produto financeiro com seus atributos principais. */
-public record Produto(Integer id,String nome,String tipo,Double rentabilidade,String risco) {
+public record Produto(Long id,String nome,String tipo,Double rentabilidade,String risco) {
 
 }
