+++ src/main/java/br/com/mattos/simuladorinvestimento/domain/model/SimulacaoEntrada.java
@@ -1,6 +1,5 @@
 package br.com.mattos.simuladorinvestimento.domain.model;
 
-/** Dados de entrada para criar uma simulação de investimento. */
-public record SimulacaoEntrada(Integer clienteId,Double valor,Integer prazoMeses,String tipoProduto) {
+public record SimulacaoEntrada(Long clienteId,Double valor,Integer prazoMeses,String tipoProduto) {
 
 }
